# api.yaml
openapi: 3.0.0
info:
  title: Telemetry API
  version: 1.0.0
paths:
  /devices/{device_id}/telemetry/latest:
    get:
      summary: Получение последних данных телеметрии
      description: Возвращает последнее полученное значение телеметрии для устройства.
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetryData'
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /devices/{device_id}/telemetry:
    get:
      summary: Получение исторических данных телеметрии
      description: Возвращает исторические данные телеметрии для устройства за определённый период времени.
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetryData'
        '404':
          description: Данные по устройству не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Добавление данных телеметрии
      description: Позволяет добавить данные телеметрии для устройства.
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceType:
                  type: string
                createdAt:
                  type: string
                  format: date-time
                telemetryData:
                  type: object
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelemetryData'
        '404':
          description: Устройство не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    TelemetryData:
      type: object
      properties:
        deviceId:
          type: string
        deviceType:
          type: string
        createdAt:
          type: string
          format: date-time
        telemetryData:
          type: object
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string