# api/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"time"
)

type DevicesDeviceIDTelemetryGetInternalServerError ErrorResponse

func (*DevicesDeviceIDTelemetryGetInternalServerError) devicesDeviceIDTelemetryGetRes() {}

type DevicesDeviceIDTelemetryGetNotFound ErrorResponse

func (*DevicesDeviceIDTelemetryGetNotFound) devicesDeviceIDTelemetryGetRes() {}

type DevicesDeviceIDTelemetryGetOKApplicationJSON []TelemetryData

func (*DevicesDeviceIDTelemetryGetOKApplicationJSON) devicesDeviceIDTelemetryGetRes() {}

type DevicesDeviceIDTelemetryLatestGetInternalServerError ErrorResponse

func (*DevicesDeviceIDTelemetryLatestGetInternalServerError) devicesDeviceIDTelemetryLatestGetRes() {}

type DevicesDeviceIDTelemetryLatestGetNotFound ErrorResponse

func (*DevicesDeviceIDTelemetryLatestGetNotFound) devicesDeviceIDTelemetryLatestGetRes() {}

type DevicesDeviceIDTelemetryPostInternalServerError ErrorResponse

func (*DevicesDeviceIDTelemetryPostInternalServerError) devicesDeviceIDTelemetryPostRes() {}

type DevicesDeviceIDTelemetryPostNotFound ErrorResponse

func (*DevicesDeviceIDTelemetryPostNotFound) devicesDeviceIDTelemetryPostRes() {}

type DevicesDeviceIDTelemetryPostReq struct {
	DeviceType    OptString                                     `json:"deviceType"`
	CreatedAt     OptDateTime                                   `json:"createdAt"`
	TelemetryData *DevicesDeviceIDTelemetryPostReqTelemetryData `json:"telemetryData"`
}

// GetDeviceType returns the value of DeviceType.
func (s *DevicesDeviceIDTelemetryPostReq) GetDeviceType() OptString {
	return s.DeviceType
}

// GetCreatedAt returns the value of CreatedAt.
func (s *DevicesDeviceIDTelemetryPostReq) GetCreatedAt() OptDateTime {
	return s.CreatedAt
}

// GetTelemetryData returns the value of TelemetryData.
func (s *DevicesDeviceIDTelemetryPostReq) GetTelemetryData() *DevicesDeviceIDTelemetryPostReqTelemetryData {
	return s.TelemetryData
}

// SetDeviceType sets the value of DeviceType.
func (s *DevicesDeviceIDTelemetryPostReq) SetDeviceType(val OptString) {
	s.DeviceType = val
}

// SetCreatedAt sets the value of CreatedAt.
func (s *DevicesDeviceIDTelemetryPostReq) SetCreatedAt(val OptDateTime) {
	s.CreatedAt = val
}

// SetTelemetryData sets the value of TelemetryData.
func (s *DevicesDeviceIDTelemetryPostReq) SetTelemetryData(val *DevicesDeviceIDTelemetryPostReqTelemetryData) {
	s.TelemetryData = val
}

type DevicesDeviceIDTelemetryPostReqTelemetryData struct{}

// Ref: #/components/schemas/ErrorResponse
type ErrorResponse struct {
	Code    OptInt    `json:"code"`
	Message OptString `json:"message"`
}

// GetCode returns the value of Code.
func (s *ErrorResponse) GetCode() OptInt {
	return s.Code
}

// GetMessage returns the value of Message.
func (s *ErrorResponse) GetMessage() OptString {
	return s.Message
}

// SetCode sets the value of Code.
func (s *ErrorResponse) SetCode(val OptInt) {
	s.Code = val
}

// SetMessage sets the value of Message.
func (s *ErrorResponse) SetMessage(val OptString) {
	s.Message = val
}

// NewOptDateTime returns new OptDateTime with value set to v.
func NewOptDateTime(v time.Time) OptDateTime {
	return OptDateTime{
		Value: v,
		Set:   true,
	}
}

// OptDateTime is optional time.Time.
type OptDateTime struct {
	Value time.Time
	Set   bool
}

// IsSet returns true if OptDateTime was set.
func (o OptDateTime) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptDateTime) Reset() {
	var v time.Time
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptDateTime) SetTo(v time.Time) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptDateTime) Get() (v time.Time, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptDateTime) Or(d time.Time) time.Time {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// NewOptInt returns new OptInt with value set to v.
func NewOptInt(v int) OptInt {
	return OptInt{
		Value: v,
		Set:   true,
	}
}

// OptInt is optional int.
type OptInt struct {
	Value int
	Set   bool
}

// IsSet returns true if OptInt was set.
func (o OptInt) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptInt) Reset() {
	var v int
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptInt) SetTo(v int) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptInt) Get() (v int, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptInt) Or(d int) int {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/components/schemas/TelemetryData
type TelemetryData struct {
	DeviceId      OptString                   `json:"deviceId"`
	DeviceType    OptString                   `json:"deviceType"`
	CreatedAt     OptDateTime                 `json:"createdAt"`
	TelemetryData *TelemetryDataTelemetryData `json:"telemetryData"`
}

// GetDeviceId returns the value of DeviceId.
func (s *TelemetryData) GetDeviceId() OptString {
	return s.DeviceId
}

// GetDeviceType returns the value of DeviceType.
func (s *TelemetryData) GetDeviceType() OptString {
	return s.DeviceType
}

// GetCreatedAt returns the value of CreatedAt.
func (s *TelemetryData) GetCreatedAt() OptDateTime {
	return s.CreatedAt
}

// GetTelemetryData returns the value of TelemetryData.
func (s *TelemetryData) GetTelemetryData() *TelemetryDataTelemetryData {
	return s.TelemetryData
}

// SetDeviceId sets the value of DeviceId.
func (s *TelemetryData) SetDeviceId(val OptString) {
	s.DeviceId = val
}

// SetDeviceType sets the value of DeviceType.
func (s *TelemetryData) SetDeviceType(val OptString) {
	s.DeviceType = val
}

// SetCreatedAt sets the value of CreatedAt.
func (s *TelemetryData) SetCreatedAt(val OptDateTime) {
	s.CreatedAt = val
}

// SetTelemetryData sets the value of TelemetryData.
func (s *TelemetryData) SetTelemetryData(val *TelemetryDataTelemetryData) {
	s.TelemetryData = val
}

func (*TelemetryData) devicesDeviceIDTelemetryLatestGetRes() {}
func (*TelemetryData) devicesDeviceIDTelemetryPostRes()      {}

type TelemetryDataTelemetryData struct{}